refactoring

+++ src/main/java/org/appMain/config/AppConfig.java
@@ -14,6 +14,7 @@ import org.springframework.security.config.annotation.method.configuration.Enabl
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.util.Arrays;
 import java.util.Calendar;
@@ -48,15 +49,15 @@ public class AppConfig {
             Supplier supplier1 = new Supplier("Petrov", "Co company");
             Supplier supplier2 = new Supplier("Ivanov", "SupplyCom");
 
-            Product p1 = new Product("water", 20d, new Date(2133333423423L), supplier1);
-            Product p2 = new Product("apple", 12d, new Date(213333), supplier2);
-            Product p3 = new Product("orange", 34d, new Date(2133333423423L), supplier1);
-            Product p4 = new Product("cookie", 54d, new Date(3333312423423L), supplier2);
-            Product p5 = new Product("pill", 212d, new Date(2123333423423L), supplier2);
-            Product p6 = new Product("bread", 2d, new Date(3121333333333L), supplier2);
-            Product p7 = new Product("cola", 1111d, new Date(1133333333213L), supplier1);
-            Product p8 = new Product("sprite", 167d, new Date(1133333333213L), supplier1);
-            Product p9 = new Product("fanta", 1d, new Date(3133333333213L), supplier1);
+            Product p1 = new Product("water", BigDecimal.valueOf(20d), new Date(2133333423423L), supplier1);
+            Product p2 = new Product("apple", BigDecimal.valueOf(12d), new Date(213333), supplier2);
+            Product p3 = new Product("orange", BigDecimal.valueOf(34d), new Date(2133333423423L), supplier1);
+            Product p4 = new Product("cookie", BigDecimal.valueOf(54d), new Date(3333312423423L), supplier2);
+            Product p5 = new Product("pill", BigDecimal.valueOf(212d), new Date(2123333423423L), supplier2);
+            Product p6 = new Product("bread", BigDecimal.valueOf(2d), new Date(3121333333333L), supplier2);
+            Product p7 = new Product("cola", BigDecimal.valueOf(1111d), new Date(1133333333213L), supplier1);
+            Product p8 = new Product("sprite", BigDecimal.valueOf(167d), new Date(1133333333213L), supplier1);
+            Product p9 = new Product("fanta", BigDecimal.valueOf(1d), new Date(3133333333213L), supplier1);
 
             Calendar calendar = Calendar.getInstance();
             java.util.Date currentDate = calendar.getTime();
